Error codes uniformisation
- definition of OPENSSL return codes and CX return codes (convention
differs)
- replacement of many'magical numbers'0 and 1 return values by return
 codes,
- other magical values suppressed
- move bn_t definition outside of ec.h (child class defining mother)
- externalisation of sys_os_perso_derive_node_with_seed_key for starknet
need
