+++ src/bolos/os_bip32.h
@@ -18,3 +18,8 @@ void expand_seed_bip32(const cx_curve_domain_t *domain, uint8_t *seed,
                        unsigned int seed_length, extended_private_key *key);
 int unhex(uint8_t *dst, size_t dst_size, const char *src, size_t src_size);
 size_t get_seed_from_env(const char *name, uint8_t *seed, size_t max_size);
+
+unsigned long sys_os_perso_derive_node_with_seed_key(
+    unsigned int mode, cx_curve_t curve, const unsigned int *path,
+    unsigned int pathLength, unsigned char *privateKey, unsigned char *chain,
+    unsigned char *seed_key, unsigned int seed_key_length);
