+++ src/bolos/os_eip2333.c
@@ -1,8 +1,10 @@
 #include <err.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
+#include "cx_bn.h"
 #include "cx_ec.h"
 #include "cx_hash.h"
 #include "cx_hkdf.h"
